% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{chisq_pval}
\alias{chisq_pval}
\title{P-value of chi-squared contingency table test}
\usage{
chisq_pval(y, g, pvdigits = 4)
}
\arguments{
\item{y}{A numeric vector (or factor). Row variable in contingency table.}

\item{g}{A numeric vector (or factor). Column variable in contingency table.}

\item{pvdigits}{number of digits for the p-value}
}
\value{
p-value of the chi squared contingency table test chisq.test(table(y,g))
}
\description{
P-value of chi-squared contingency table test
}
\details{
If conditions for the validity of chi squared test are not met (expected
frequencies less than 5) a Fisher exact test is performed instead.
}
\examples{
df <- data.frame(g=rbinom(100,4,0.5),y=rnorm(sample(c(1:2),100,replace=TRUE)))
chisq_pval(df$y,df$g,pvdigits=3)
}
