\name{custHeaderPanel}
\alias{custHeaderPanel}
\alias{renderRepTools}
\title{Helper functions for using ReportingTools with Shiny}
\description{
  \code{custHeaderPanel} defines a header panel with custom Javascript and CSS
  included.

  \code{renderRepTools} declares to shiny that the output is raw HTML
  that was formatted by ReportingTools.
  
  See the shiny vignette for example usage.
}
\author{
  Gabriel Becker and Jessica L. Larson 
}