% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{describe}
\alias{describe}
\title{Describe a variable depending on normality}
\usage{
describe(x, digits = 2, probs = c(0.25, 0.75), alpha = 0.05, na.rm = TRUE)
}
\arguments{
\item{x}{variable to be described.}

\item{digits}{decimal digits in the result.}

\item{probs}{quantiles to be used in case of non normality.}

\item{alpha}{significance level for testing normality (0.05 by default).}

\item{na.rm}{Should missing values be considered?. If TRUE and there are missing values present, the funcion returns NA.}
}
\value{
Mean ± sd of variable if normal or Median and Interquartile range if not normal.
}
\description{
Describe a variable depending on normality
}
\examples{
describe(rnorm(100))
describe(runif(100))
}
