\name{filename-methods}
\docType{methods}
\alias{filename}
\alias{filename-methods}
\alias{filename,CSVFile-method}
\alias{filename,HTMLReport-method}
\title{ 
Methods for getting the name of a file for a CSVFile or HTMLReport.
}
\description{
These methods return the filename for a report, for report types where a single file is produced.
}
\section{Methods}{
\describe{

\item{\code{signature(object = "CSVFile")}}{
Return the file name of the CSVFile. This is generated by adding '.csv' to the shortName of the report. 
}

\item{\code{signature(object = "HTMLReport")}}{
Return the file name of the HTMLReport. This is generated by adding '.html' to the shortName of the report. 
}
}}
\keyword{methods}
