% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{formatPval}
\alias{formatPval}
\title{Format p-values}
\usage{
formatPval(x, pvdigits = 4)
}
\arguments{
\item{x}{p-value to be formatted.}

\item{pvdigits}{number of digits in the result.}
}
\value{
p-value rounded to pvdigits decimal digits, or an expression like "<0.0001" when x<1e-pvdigits.
}
\description{
Format p-values
}
\examples{
formatPval(0.0000001,pvdigits=3)
formatPval(0.345251,pvdigits=4)
}
