% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{isNormal}
\alias{isNormal}
\title{Testing if a variable follows a normal distribution using Shapiro-Wilk test.}
\usage{
isNormal(x, alpha = 0.05)
}
\arguments{
\item{x}{Variable whose normality is to be tested.}

\item{alpha}{Significance level for testing normality.}
}
\value{
TRUE (if normality can be accepted) or FALSE (if normality is rejected).
}
\description{
Testing if a variable follows a normal distribution using Shapiro-Wilk test.
}
\examples{
isNormal(rnorm(100))
isNormal(runif(100))
}
