% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{kruskal_pval}
\alias{kruskal_pval}
\title{P-value of Kruskal-Wallis test}
\usage{
kruskal_pval(y, g, pvdigits = 4)
}
\arguments{
\item{y}{Response variable.}

\item{g}{Grouping variable.}

\item{pvdigits}{number of digits for the p-value.}
}
\value{
p-value of the analysis of the Kruskal-Wallis test kruskal.test(y~g).
}
\description{
P-value of Kruskal-Wallis test
}
\examples{
df <- data.frame(g=gl(3,10),y=runif(30))
kruskal_pval(df$y,df$g,pvdigits=3)
}
