% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{kruskal_Wilcox_pval}
\alias{kruskal_Wilcox_pval}
\title{P-value of Wilcoxon of Kruskal-Wallis test}
\usage{
kruskal_Wilcox_pval(y, g, pvdigits = 4)
}
\arguments{
\item{y}{Response variable.}

\item{g}{Grouping variable.}

\item{pvdigits}{number of digits for the p-value.}
}
\value{
p-value of Wilcoxon test (when comparing two groups) or Kruskal-Wallis test
(when comparing more than two groups).
}
\description{
P-value of Wilcoxon of Kruskal-Wallis test
}
\examples{
df <- data.frame(g=gl(3,10),y=runif(30))
kruskal_Wilcox_pval(df$y,df$g,pvdigits=3)
}
