% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{meanSd}
\alias{meanSd}
\title{Arithmetic mean ± sd}
\usage{
meanSd(x, digits = 2, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{digits}{Number of decimal digits for showing the mean and sd of x.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
A string with the value of mean ± sd of the vector x.
}
\description{
Arithmetic mean ± sd
}
\examples{
meanSd(rnorm(50))

}
