% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{medianIQR}
\alias{medianIQR}
\title{Median and Interquartile Range}
\usage{
medianIQR(x, digits = 2, probs = c(0.25, 0.75), roundFrom = 30, na.rm = TRUE)
}
\arguments{
\item{x}{A vector of numeric values.}

\item{digits}{Number of decimal digits for showing the median and Interquartile Range of x.}

\item{probs}{Pair of quantiles to be computed around the median. Default are q25 and q75.}

\item{roundFrom}{If median is greater than this value, median and IQR are rounded to zero decimals even if digits>0.}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
A character with the value of the median and Interquartile Range.
}
\description{
Median and Interquartile Range
}
\examples{
medianIQR(rnorm(50))
medianIQR(runif(500,0,3))

}
