\name{mockRnaSeqData}
\alias{mockRnaSeqData}
\docType{data}
\title{
A counts table of mock RNA-deq data in mouse. 
}
\description{
Expression data from an RNA-seq experiment.  Rows were randomly assigned mouse Entrez ids.
}
\usage{data(mockRnaSeqData)}
\examples{
	data(mockRnaSeqData)
}
\keyword{datasets}
