% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{nPctBin01}
\alias{nPctBin01}
\title{Absolute and relative frequencies of 1's in a binary 0-1 variable.}
\usage{
nPctBin01(x, digits = 1)
}
\arguments{
\item{x}{A vector of numeric values containing only 1's and 0's. Function returns an error if there are other values in x.}

\item{digits}{Number of decimal digits for showing relative frequency.}
}
\value{
Number and percent of 1's in a numeric vector x which contains only 1's and 0's.
}
\description{
Absolute and relative frequencies of 1's in a binary 0-1 variable.
}
\examples{
nPctBin01(rbinom(100,1,0.4))
nPctBin01(sample(c(0,1),50,replace=TRUE))
}
