\name{renderRepTools}
\alias{renderRepTools}
\title{Tells shiny how to use ReportingTools for server.R}
\usage{
    custHeaderPanel(expr, env=parent.frame(), quoted=FALSE) 
}
\arguments{
  \item{expr}{}

  \item{env}{}

  \item{quoted}{}

}

\value{
  

}
\description{
	
}
\details{
	
}
\author{
  Gabriel Becker and Jessica L. Larson 
}