% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{report_format}
\alias{report_format}
\title{Elegant format for report tables}
\usage{
report_format(report_data, caption = NULL, fontsize = NULL)
}
\arguments{
\item{report_data}{report_data Report table to be formatted.}
}
\value{
Table formatted with flextable.
}
\description{
Elegant format for report tables
}
\examples{
df <- data.frame(x=rbinom(90,3,0.6),y=rbinom(90,4,0.8), z=rbinom(90,5,0.5),
g=sample(c("Yes","No"),90,replace=TRUE))
df \%>\%
report_nPct(c(x,y,z)) \%>\% report_format()  # Overall summary of variables x, y and z
df \%>\%
report_nPct(c(x,y,z), groupVar=g, digits=3) \%>\% report_format(caption="My report table")

}
