% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R
\name{report_medianIQR}
\alias{report_medianIQR}
\title{Report table of medians ± IQR (Interquartile Range)}
\usage{
report_medianIQR(
  data,
  summary_vars,
  groupVar = NULL,
  digits = 2,
  roundFrom = 30,
  probs = c(0.25, 0.75),
  pvdigits = 4,
  na.rm = TRUE
)
}
\arguments{
\item{data}{data frame or tibble which contains the data.}

\item{summary_vars}{Variable or variables whose median and quartiles is to be calculated.}

\item{groupVar}{Grouping variable.}

\item{digits}{Number of decimal digits for the results.}

\item{roundFrom}{If median is greater than this value, median and IQR are rounded to zero decimals even if digits>0.}

\item{probs}{Pair of quantiles to be computed around the median. Default are q25 and q75.}

\item{pvdigits}{Number of decimal digits for the p-value of Kruskal test.}

\item{na.rm}{Should NA values be removed (possible values are TRUE or FALSE)}
}
\value{
A table with the overall median and quartiles of the variables and, if a grouping
variable is specified, the medians and quartiles by group and the p-value of the kruskal test for
comparing location parameters.
}
\description{
report_medianIQR builds a table with the overall medians and (by
default) quartiles 25 and 75 of one or more variables. A grouping variable can also be
specified, in which case medians and quartiles are also calculated for each group and the
p-value of a Kruskal-Wallis test is displayed to test the null hypothesis that the location
parameters of the distribution of the variable are the same in each group.
}
\examples{
df <- data.frame(x=rnorm(100,10,3),y=rnorm(100,50,8), z=runif(100,20,30),g=sample(c("Yes","No"),100,replace=TRUE))
df \%>\%
report_medianIQR(c(x,y,z))  # Only overall medians and quartiles of variables x, y and z
df \%>\%
report_medianIQR(c(x,y,z), groupVar=g, digits=1)
}
