% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_transpose.R
\name{report_transpose}
\alias{report_transpose}
\title{Transpose report table}
\usage{
report_transpose(report_data)
}
\arguments{
\item{report_data}{table to transpose}
}
\value{
transposed table
}
\description{
Transpose report table
}
\note{
At the moment this function only works when there is only one variable in report.
}
\examples{
df <- data.frame(V1=rbinom(500,1,0.4),V2=rbinom(500,1,0.5),
   group=gl(5,100))
df \%>\% report_nPctBin01(V1,groupVar = group) \%>\%
   report_transpose()

}
