\name{custHeaderPanel}
\alias{custHeaderPanel}
\title{Tells shiny how to use ReportingTools in ui.R}
\usage{
    custHeaderPanel(title, window, windowTitle = title, js = NULL, css = NULL) 
}
\arguments{
  \item{title}{}

  \item{window}{}

  \item{windowTitle}{}

  \item{js}{}

  \item{css}{}
}

\value{
  

}
\description{
	
}
\details{
	
}
\author{
  Gabriel Becker and Jessica L. Larson 
}